[4.0.1] [+]更换等级与经验的数据类型为 Long

+++ src/main/kotlin/com/github/cpjinan/plugin/akarilevel/common/event/exp/PlayerExpChangeEvent.kt
@@ -15,7 +15,7 @@ import taboolib.platform.type.BukkitProxyEvent
 class PlayerExpChangeEvent(
     val player: Player,
     var levelGroup: String,
-    var expAmount: Int,
+    var expAmount: Long,
     var source: String
 ) :
     BukkitProxyEvent()
