+++ src/main/kotlin/com/github/cpjinan/plugin/akarilevel/internal/database/type/PlayerData.kt
@@ -3,4 +3,4 @@ package com.github.cpjinan.plugin.akarilevel.internal.database.type
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class PlayerData(var level: Int = 0, var exp: Int = 0)
+data class PlayerData(var level: Long = 0, var exp: Long = 0)
